Added support for env variables and added extra security for purging database

+++ api/src/handlers/misc.ts
@@ -6,7 +6,7 @@ import { MongoService } from '../services/mongo';
 export const MiscHandler = {
 
   getPingResponse: (_req: express.Request, res: express.Response) => {
-    const response: ISimpleResponse = { code: 'success', message: 'PONG', time: moment().unix() };
+    const response: ISimpleResponse = { code: 'success', message: 'pong', time: moment().unix() };
     res.send(response);
   },
 
@@ -14,10 +14,17 @@ export const MiscHandler = {
     res.redirect('https://documenter.getpostman.com/view/8555555/SWT5gzbe?version=latest');
   },
 
-  resetDatabase: async (_req: express.Request, res: express.Response) => {
-    await MongoService.deleteEverything();
+  resetDatabase: async (req: express.Request, res: express.Response) => {
+    // Safeguard to ensure this isn't trigerred accidentally
+    if (req.params.code !== process.env.PURGE_CODE) {
+      const response: ISimpleResponse = { code: 'failed', message: 'invalid purge code', time: moment().unix() };
+      res.send(response);
+      return false;
+    } 
 
-    const response: ISimpleResponse = { code: 'success', message: 'Purge complete', time: moment().unix() };
+    await MongoService.deleteEverything();
+    const response: ISimpleResponse = { code: 'success', message: 'purge complete', time: moment().unix() };
     res.send(response);
+    return true;
   }
 }
